+++ public/js/app.js
@@ -3,30 +3,38 @@ import {Heatmap} from "./leaflet/heatmap";
 import {DataTable, DataTableStorage} from "./data-table";
 import Router from "./router";
 import Chat from "./chat";
-import {resolveEchartsCallbacks, compareVersions} from "./utils";
+import {resolveEchartsCallbacks} from "./utils";
+import {updateGithubLatestRelease} from "./github";
+import LazyLoad from "../libraries/lazyload.min";
+import lightGallery from "../libraries/lightgallery/lightgallery.umd.min";
+import lgZoom from "../libraries/lightgallery/lightgallery.lg-zoom.min";
 
 const $main = document.querySelector("main");
 const $sideNav = document.querySelector("aside");
 const $topNav = document.querySelector("nav");
 
-// Toggle sidebar collapsed state.
-const sideNavCollapsed = localStorage.getItem('sideNavCollapsed');
-if (sideNavCollapsed === 'true') {
-    $main.classList.add('sidebar-is-collapsed');
-    $sideNav.classList.add('sidebar-is-collapsed');
-    $topNav.classList.add('sidebar-is-collapsed');
-}
-
 const resizeAllCharts = () => allRenderedCharts.forEach(chart => chart.resize());
 
-const toggleSidebar = () => {
+const collapsed = localStorage.getItem('sideNavCollapsed') === 'true';
+[$main, $sideNav, $topNav].forEach((el) =>
+    el.classList.toggle('sidebar-is-collapsed', collapsed)
+);
+
+document.getElementById('toggle-sidebar-collapsed-state').addEventListener('click', () => {
     const collapsed = $main.classList.toggle('sidebar-is-collapsed');
     [$sideNav, $topNav].forEach(el => el.classList.toggle('sidebar-is-collapsed', collapsed));
+
     localStorage.setItem('sideNavCollapsed', String(collapsed));
     resizeAllCharts();
-};
-document.getElementById('toggle-sidebar-collapsed-state')?.addEventListener('click', toggleSidebar);
+});
 
+// Enable image lazy load.
+const lazyLoad = new LazyLoad({
+    thresholds: "50px",
+    callback_error: (img) => {
+        img.setAttribute("src", window.statisticsForStrava.placeholderBrokenImage);
+    }
+});
 
 // Boot router.
 const router = new Router($main);
@@ -93,7 +101,7 @@ if ($modalAIChat) {
 
 document.addEventListener('modalWasLoaded.ai-chat', (e) => {
     const $modal = e.detail.modal;
-    Chat($modal).render();
+    new Chat($modal).render();
 });
 
 const initElements = (rootNode) => {
@@ -315,73 +323,6 @@ export const openModal = (modalId) => {
 
     modal.show();
 }
-
-const updateGithubLatestRelease = async () => {
-    const $latestVersionEl = document.querySelector('[data-latest-version]');
-    if (!$latestVersionEl) return;
-
-    const currentVersion = $latestVersionEl.dataset.currentVersion;
-    if (!currentVersion) return;
-
-    const CACHE_KEY = 'latestReleaseCache';
-    const CACHE_TTL = 1000 * 60 * 60 * 6; // 6 hours
-
-    const now = Date.now();
-    const cached = JSON.parse(localStorage.getItem(CACHE_KEY) || 'null');
-
-    // If cache exists and is recent, use it
-    if (cached && now - cached.timestamp < CACHE_TTL) {
-        if (compareVersions(currentVersion, cached.latestVersion) < 0) {
-            // Update available.
-            showLatestVersion(cached.latestVersion);
-        }
-        return;
-    }
-
-    try {
-        const releaseResponse = await fetch(
-            'https://api.github.com/repos/robiningelbrecht/statistics-for-strava/releases/latest',
-            { headers: { 'Accept': 'application/vnd.github+json' } }
-        );
-        if (!releaseResponse.ok) throw new Error('Failed to fetch latest release');
-
-        const latestVersion = (await releaseResponse.json()).name;
-
-        if (currentVersion === latestVersion) {
-            localStorage.setItem(CACHE_KEY, JSON.stringify({ latestVersion, timestamp: now }));
-            return;
-        }
-
-        // Verify Docker image workflow run success
-        const workflowRunsResponse = await fetch(
-            `https://api.github.com/repos/robiningelbrecht/statistics-for-strava/actions/runs?event=push&status=completed&conclusion=success&exclude_pull_requests=true&branch=${latestVersion}`,
-            { headers: { 'Accept': 'application/vnd.github+json' } }
-        );
-        if (!workflowRunsResponse.ok) throw new Error('Failed to fetch workflow runs');
-
-        const { workflow_runs = [] } = await workflowRunsResponse.json();
-        const dockerWorkflowRun = workflow_runs.find(
-            run => run.path === '.github/workflows/docker-image.yml'
-        );
-
-        if(!dockerWorkflowRun) return;
-
-        showLatestVersion(latestVersion);
-        localStorage.setItem(CACHE_KEY, JSON.stringify({ latestVersion, timestamp: now }));
-    } catch (err) {
-        console.error('Error checking latest release:', err);
-    }
-}
-
-const showLatestVersion = (latestVersion) => {
-    const $latestVersionEl = document.querySelector('[data-latest-version]');
-    const $link = $latestVersionEl.querySelector('a');
-    if ($link) {
-        $link.href = $link.href.replace('[LATEST_VERSION]', latestVersion);
-        $link.textContent = $link.textContent.replace('[LATEST_VERSION]', latestVersion);
-    }
-    $latestVersionEl.classList.remove('hidden');
-};
 
 (async () => {
     await updateGithubLatestRelease();
