+++ public/js/utils.js
@@ -74,19 +74,3 @@ export const resolveEchartsCallbacks = (obj, path) => {
     resolvePath(obj, parts);
 };
 
-export const compareVersions = (a, b) => {
-    a = a.replace(/^v/, '');
-    b = b.replace(/^v/, '');
-
-    const pa = a.split('.').map(Number);
-    const pb = b.split('.').map(Number);
-
-    for (let i = 0; i < Math.max(pa.length, pb.length); i++) {
-        const na = pa[i] || 0;
-        const nb = pb[i] || 0;
-        if (na > nb) return 1;
-        if (na < nb) return -1;
-    }
-    return 0;
-}
-
