+++ webpack.config.js
@@ -8,7 +8,8 @@ module.exports = {
     },
     output: {
         path: path.resolve(__dirname, 'public/js/dist'),
-        filename: '[name].js'
+        filename: '[name].js',
+        clean: true,
     },
     resolve: {
         fallback: {
